+++ lightgbm/src/test/scala/com/microsoft/azure/synapse/ml/lightgbm/split1/VerifyLightGBMClassifierStreamOnly.scala
@@ -1,15 +1,12 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.lightgbm.split4
-
-import com.microsoft.azure.synapse.ml.lightgbm.split1._
+package com.microsoft.azure.synapse.ml.lightgbm.split1
 
 // scalastyle:off magic.number
 /** Tests to validate the functionality of LightGBM module in streaming mode. */
 class VerifyLightGBMClassifierStreamOnly extends LightGBMClassifierTestData {
   override def ignoreSerializationFuzzing: Boolean = true
-
   override def ignoreExperimentFuzzing: Boolean = true
 
   test("Verify LightGBMClassifier handles global sample mode correctly") {
@@ -43,6 +40,4 @@ class VerifyLightGBMClassifierStreamOnly extends LightGBMClassifierTestData {
       fitModel.transform(df)
     }
   }
-
-
 }
