+++ lightgbm/src/test/scala/com/microsoft/azure/synapse/ml/lightgbm/split3/VerifyLightGBMClassifierStreamBasic.scala
@@ -1,9 +1,10 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.lightgbm.split1
+package com.microsoft.azure.synapse.ml.lightgbm.split3
 
-import com.microsoft.azure.synapse.ml.lightgbm._
+import com.microsoft.azure.synapse.ml.lightgbm.{LightGBMClassificationModel, LightGBMConstants}
+import com.microsoft.azure.synapse.ml.lightgbm.split1._
 import org.apache.spark.ml.tuning.{ParamGridBuilder, TrainValidationSplit}
 
 // scalastyle:off magic.number
