+++ lightgbm/src/test/scala/com/microsoft/azure/synapse/ml/lightgbm/split4/VerifyLightGBMClassifierBulkBasic.scala
@@ -1,10 +1,10 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.lightgbm.split3
+package com.microsoft.azure.synapse.ml.lightgbm.split4
 
 import com.microsoft.azure.synapse.ml.lightgbm._
-import com.microsoft.azure.synapse.ml.lightgbm.split1.VerifyLightGBMClassifierStreamBasic
+import com.microsoft.azure.synapse.ml.lightgbm.split3.VerifyLightGBMClassifierStreamBasic
 
 /** Tests to validate the functionality of LightGBM module. */
 class VerifyLightGBMClassifierBulkBasic extends VerifyLightGBMClassifierStreamBasic {
