+++ lightgbm/src/test/scala/com/microsoft/azure/synapse/ml/lightgbm/split6/VerifyLightGBMClassifierBulk.scala
@@ -1,10 +1,10 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.lightgbm.split5
+package com.microsoft.azure.synapse.ml.lightgbm.split6
 
 import com.microsoft.azure.synapse.ml.lightgbm._
-import com.microsoft.azure.synapse.ml.lightgbm.split4.VerifyLightGBMClassifierStream
+import com.microsoft.azure.synapse.ml.lightgbm.split5.VerifyLightGBMClassifierStream
 
 /** Tests to validate the functionality of LightGBM module. */
 class VerifyLightGBMClassifierBulk extends VerifyLightGBMClassifierStream {
